Avoid using TryFinallyWithYield OpCode when there is no Yield. Implement partial tracking of Yield within ByteCodeGenerator to enable this.
The OpCode TryFinallyWithYield is not supported by the Jit. It was being used unnecessarily in various code patterns that should be safe to Jit, such patterns were hitting an Abort in the Jit because of the unimplemented OpCode.

As the OpCode is unnecessary when there is no Yield (the cases we currently try to jit) fix logic so we do not use it in those cases.